+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Carts/Commands/UpsertCart/UpsertCartProfile.cs
@@ -1,14 +1,14 @@
 using Ambev.DeveloperEvaluation.Application.Carts.Commands;
 using AutoMapper;
 
-namespace Ambev.DeveloperEvaluation.WebApi.Features.Carts.Commands.CreateCart;
+namespace Ambev.DeveloperEvaluation.WebApi.Features.Carts.Commands.UpsertCart;
 
-public sealed class CreateCartProfile : Profile
+public sealed class UpsertCartProfile : Profile
 {
-    public CreateCartProfile()
+    public UpsertCartProfile()
     {
         // In
-        CreateMap<CreateCartRequest, CreateCartCommand>()
+        CreateMap<UpsertCartRequest, UpsertCartCommand>()
             .ForMember(dest => dest.UserId, opt => opt.MapFrom(src => src.UserId))
             .ForMember(dest => dest.Date, opt => opt.MapFrom(src => src.Date))
             .ForMember(dest => dest.Products, opt => opt.MapFrom(src => src.Products));
