+++ src/Ambev.DeveloperEvaluation.WebApi/Features/Carts/Commands/UpsertCart/UpsertCartRequest.cs
@@ -1,8 +1,8 @@
 using Ambev.DeveloperEvaluation.WebApi.Features.Carts.Common;
 
-namespace Ambev.DeveloperEvaluation.WebApi.Features.Carts.Commands.CreateCart;
+namespace Ambev.DeveloperEvaluation.WebApi.Features.Carts.Commands.UpsertCart;
 
-public sealed class CreateCartRequest
+public sealed class UpsertCartRequest
 {
     public Guid UserId { get; set; }    
     public required DateTime Date { get; set; }
